Add some logic to make the Yes/No questions work
Add logic to main.js to support Yes/No question to be evaluated according to the
desired answer

Add logic to main.js to support template selection and configuration based
upon the questions.json file. The idea here is that the configuring user
may use an existing template or may add an additional template to support
their question/answer scenarios.

Add a section called non_question_responses to questions.json to provide
a lookup for unqualified_person template name. IMO this section may be
leveraged further to provide custom responses in scenarios where that is
required.

Change questions.json to reflect the problem space a bit better. (IMO)

Add a template for unqualified_person that shows a message that says user
may not be qualified and provides a link to the qualifications for
Oklahoma.